worker/requirements.txt
boto3==1.26.34
textract-trp==0.1.3
pymongo==3.12.0
python-dotenv==0.21.0
rq==1.11.1
pottery==3.0.0
PyPDF2==2.12.1
